feat(access): new 'Access on Swarm' sidebar page for hash-retrieve flow
Replaces the dead-code RetrieveModal in Drive.tsx with a dedicated
page at /access. Adds a sidebar entry (Download icon) labeled
'Access on Swarm'. Same fields as before (Swarm hash + optional
filename), same behavior (downloadFile with fallback to
downloadBytes), but now reachable and treated as a first-class
'reach any content on the network' surface.

Cleanup: removed RetrieveModal function, retrieveOpen state, and
the two render sites in Drive.tsx — all dead code since we removed
the Retrieve button from the top bar earlier.
