+++ audit.json
@@ -46,6 +46,16 @@
             "date": 1776186749871,
             "name": "/Users/gzupan/Library/Logs/Nook/bee.7.log",
             "hash": "566c2afe283acc7db6a630c2ab1fd92a"
+        },
+        {
+            "date": 1779890360448,
+            "name": "/Users/gzupan/Library/Logs/Nook/bee.8.log",
+            "hash": "3b4f8f7edd23d002f1a1dd2b708f4884"
+        },
+        {
+            "date": 1779907670467,
+            "name": "/Users/gzupan/Library/Logs/Nook/bee.9.log",
+            "hash": "e59d5d7a5482d4ba41ca9d75ec6bc468"
         }
     ]
 }
