+++ ui/src/App.tsx
@@ -1,5 +1,6 @@
 import { Navigate, Route, Routes } from 'react-router-dom'
 import Layout from './components/Layout'
+import AccessOnSwarm from './pages/AccessOnSwarm'
 import Account from './pages/Account'
 import Contacts from './pages/Contacts'
 import Dev from './pages/Dev'
@@ -17,6 +18,7 @@ export default function App() {
         <Route path="wallet" element={<Navigate to="/account" replace />} />
         <Route path="account" element={<Account />} />
         <Route path="contacts" element={<Contacts />} />
+        <Route path="access" element={<AccessOnSwarm />} />
         <Route path="settings" element={<Settings />} />
         <Route path="logs" element={<Navigate to="/dev" replace />} />
         <Route path="dev" element={<Dev />} />
