+++ ui/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/theme.ts","./src/vite-env.d.ts","./src/wagmi.ts","./src/api/bee.test.ts","./src/api/bee.ts","./src/api/client.ts","./src/api/feeds.ts","./src/api/queries.ts","./src/api/server.ts","./src/apps/websitepublisher.tsx","./src/components/addshareddrivemodal.tsx","./src/components/ensmodal.tsx","./src/components/layout.tsx","./src/components/onboarding.tsx","./src/components/sharemodal.tsx","./src/components/walletgate.tsx","./src/crypto/signer.ts","./src/hooks/usederivedkey.ts","./src/hooks/usedrivemetadata.ts","./src/hooks/useshareddrives.ts","./src/hooks/useupload.ts","./src/hooks/useuploadhistory.ts","./src/pages/account.tsx","./src/pages/dev.tsx","./src/pages/drive.tsx","./src/pages/logs.tsx","./src/pages/overview.tsx","./src/pages/settings.tsx","./src/pages/wallet.tsx","./src/store/app.ts","./src/store/identity.ts","./src/utils/directory.ts","./src/utils/tar.ts","./vite.config.ts"],"version":"5.9.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/theme.ts","./src/vite-env.d.ts","./src/wagmi.ts","./src/api/bee.test.ts","./src/api/bee.ts","./src/api/client.ts","./src/api/feeds.ts","./src/api/queries.ts","./src/api/server.ts","./src/apps/messages.tsx","./src/apps/websitepublisher.tsx","./src/components/addshareddrivemodal.tsx","./src/components/ensmodal.tsx","./src/components/layout.tsx","./src/components/onboarding.tsx","./src/components/sharemodal.tsx","./src/components/swarmnotifytest.tsx","./src/components/walletgate.tsx","./src/components/ui/badge.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/input.tsx","./src/components/ui/sidebar.tsx","./src/components/ui/switch.tsx","./src/components/ui/tabs.tsx","./src/components/ui/textarea.tsx","./src/crypto/signer.ts","./src/hooks/usederivedkey.ts","./src/hooks/usedrivemetadata.ts","./src/hooks/useinboxpolling.ts","./src/hooks/useregistrypolling.ts","./src/hooks/useshareddrives.ts","./src/hooks/useupload.ts","./src/hooks/useuploadhistory.ts","./src/lib/utils.ts","./src/notify/constants.ts","./src/notify/contact-state.ts","./src/notify/invitations.ts","./src/notify/messages.ts","./src/notify/provider.ts","./src/notify/share-link.ts","./src/notify/storage.ts","./src/notify/types.ts","./src/pages/account.tsx","./src/pages/contacts.tsx","./src/pages/dev.tsx","./src/pages/drive.tsx","./src/pages/identity.tsx","./src/pages/logs.tsx","./src/pages/overview.tsx","./src/pages/settings.tsx","./src/pages/wallet.tsx","./src/store/app.ts","./src/store/identity.ts","./src/utils/directory.ts","./src/utils/tar.ts","./vite.config.ts"],"version":"5.9.3"}
